Improve viewport for iOS devices
With this commit, the Snake game now looks awesome on iPhone X without display issues.

On the iPad Air 2 (possibly other iPads as well), it generally works well, however, the game canvas doesn't use up all available space in landscape orientation.

On older iPhones (7 Plus, 8), the game looks good when first opened in default portrait, however, on rotate the top padding gets messed up for some reason.
This requires further work and is potentially an iOS 11.2 bug. Might have to provide a hacky workaround for this in the future.